// backend/src/types/auth.ts
import type { UserRole, UserStatus } from "../models/User.js";

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshTokenInput {
  refreshToken: string;
}

export interface AuthUserResponse {
  id: string;
  email: string;
  name: string;
  phone?: string;
  avatar?: string;
  role: UserRole;
  status: UserStatus;
  isActive: boolean;
}

export interface RegisterResponseData {
  user: AuthUserResponse;
}

export interface LoginResponseData {
  user: AuthUserResponse;
  accessToken: string;
  refreshToken: string;
}

export interface RefreshTokenResponseData {
  accessToken: string;
  refreshToken: string;
}

export interface LogoutResponseData {
  message: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  newPassword: string;
}

export interface PasswordResetResponseData {
  message: string;
}